+++ internal/dispatcher/dispatcher_test.go
@@ -24,7 +24,7 @@ func TestDispatcherSuite(t *testing.T) {
 }
 
 func (s *DispatcherSuite) SetupTest() {
-	db, err := storage.NewPebble(config.Pebble{DataPath: ""}, zap.NewNop())
+	db, err := storage.NewPebble(config.Pebble{Mode: "memory"}, zap.NewNop())
 	s.Require().NoError(err)
 	s.db = db
 }